# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'What to watch REST API'
  description: "Данный проект представляет собой REST API для управления базой данных фильмов,\n                        включая функции оставления комментариев, регистрации пользователей и управления\n                        списком избранных фильмов. Сервис интегрируется с КиноПоиском, что позволяет\n                        получать комментарии к фильмам и предоставляет пользователям возможность активно взаимодействовать с ними."
  version: 1.0.0
servers:
  -
    url: 'https://localhost'
paths:
...
  /api/v1/films:
    get:
      summary: 'GET api/v1/films'
      operationId: gETApiv1films
      description: 'Список фильмов'
      parameters:
        -
          in: query
          name: status
          description: "Администраторы могут фильтровать фильмы по разным статусам\n        У обычных пользователей этот фильтр не работает"
          example: ready
          required: false
          schema:
            type: string
            description: "Администраторы могут фильтровать фильмы по разным статусам\n        У обычных пользователей этот фильтр не работает"
            example: ready
        -
          in: query
          name: genre
          description: 'Фильтр по жанру фильма'
          example: Шутер
          required: false
          schema:
            type: string
            description: 'Фильтр по жанру фильма'
            example: Шутер
        -
          in: query
          name: orderBy
          description: 'Фильтр по атрибуту. По умолчанию сортируется по released'
          example: released
          required: false
          schema:
            type: string
            description: 'Фильтр по атрибуту. По умолчанию сортируется по released'
            example: released
        -
          in: query
          name: orderTo
          description: 'Направление фильтра по атрибуту'
          example: asc
          required: false
          schema:
            type: string
            description: 'Направление фильтра по атрибуту'
            example: asc
            enum:
              - asc
              - desc
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    -
                      id: 65
                      name: 'Example Movie 1'
                      posterImage: 'https://example.com/poster1.jpg'
                      previewImage: 'https://example.com/preview1.jpg'
                      backgroundImage: 'https://example.com/background1.jpg'
                      backgroundColor: '#FFFFFF'
                      videoLink: 'https://example.com/video1.mp4'
                      previewVideoLink: 'https://example.com/previewVideo1.mp4'
                      description: 'A great movie about...'
                      rating: '5.50'
                      scoresCount: 2
                      director: 'Director 1'
                      starring:
                        - 'Actor 1'
                        - 'Actor 2'
                      runTime: 120
                      genre:
                        - action
                      released: '2021'
                      status: ready
                      is_favorite: false
                    -
                      id: 65
                      name: 'Example Movie 1'
                      posterImage: 'https://example.com/poster1.jpg'
                      previewImage: 'https://example.com/preview1.jpg'
                      backgroundImage: 'https://example.com/background1.jpg'
                      backgroundColor: '#FFFFFF'
                      videoLink: 'https://example.com/video1.mp4'
                      previewVideoLink: 'https://example.com/previewVideo1.mp4'
                      description: 'A great movie about...'
                      rating: '5.50'
                      scoresCount: 2
                      director: 'Director 1'
                      starring:
                        - 'Actor 1'
                        - 'Actor 2'
                      runTime: 120
                      genre:
                        - action
                      released: '2021'
                      status: ready
                      is_favorite: false
                  links:
                    first: '/?page=1'
                    last: '/?page=1'
                    prev: null
                    next: null
                  meta:
                    current_page: 1
                    from: 1
                    last_page: 1
                    links:
                      -
                        url: null
                        label: pagination.previous
                        active: false
                      -
                        url: '/?page=1'
                        label: '1'
                        active: true
                      -
                        url: null
                        label: pagination.next
                        active: false
                    path: /
                    per_page: 8
                    to: 2
                    total: 2
                properties:
                  data:
                    type: array
                    example:
                      -
                        id: 65
                        name: 'Example Movie 1'
                        posterImage: 'https://example.com/poster1.jpg'
                        previewImage: 'https://example.com/preview1.jpg'
                        backgroundImage: 'https://example.com/background1.jpg'
                        backgroundColor: '#FFFFFF'
                        videoLink: 'https://example.com/video1.mp4'
                        previewVideoLink: 'https://example.com/previewVideo1.mp4'
                        description: 'A great movie about...'
                        rating: '5.50'
                        scoresCount: 2
                        director: 'Director 1'
                        starring:
                          - 'Actor 1'
                          - 'Actor 2'
                        runTime: 120
                        genre:
                          - action
                        released: '2021'
                        status: ready
                        is_favorite: false
                      -
                        id: 65
                        name: 'Example Movie 1'
                        posterImage: 'https://example.com/poster1.jpg'
                        previewImage: 'https://example.com/preview1.jpg'
                        backgroundImage: 'https://example.com/background1.jpg'
                        backgroundColor: '#FFFFFF'
                        videoLink: 'https://example.com/video1.mp4'
                        previewVideoLink: 'https://example.com/previewVideo1.mp4'
                        description: 'A great movie about...'
                        rating: '5.50'
                        scoresCount: 2
                        director: 'Director 1'
                        starring:
                          - 'Actor 1'
                          - 'Actor 2'
                        runTime: 120
                        genre:
                          - action
                        released: '2021'
                        status: ready
                        is_favorite: false
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 65
                        name:
                          type: string
                          example: 'Example Movie 1'
                        posterImage:
                          type: string
                          example: 'https://example.com/poster1.jpg'
                        previewImage:
                          type: string
                          example: 'https://example.com/preview1.jpg'
                        backgroundImage:
                          type: string
                          example: 'https://example.com/background1.jpg'
                        backgroundColor:
                          type: string
                          example: '#FFFFFF'
                        videoLink:
                          type: string
                          example: 'https://example.com/video1.mp4'
                        previewVideoLink:
                          type: string
                          example: 'https://example.com/previewVideo1.mp4'
                        description:
                          type: string
                          example: 'A great movie about...'
                        rating:
                          type: string
                          example: '5.50'
                        scoresCount:
                          type: integer
                          example: 2
                        director:
                          type: string
                          example: 'Director 1'
                        starring:
                          type: array
                          example:
                            - 'Actor 1'
                            - 'Actor 2'
                          items:
                            type: string
                        runTime:
                          type: integer
                          example: 120
                        genre:
                          type: array
                          example:
                            - action
                          items:
                            type: string
                        released:
                          type: string
                          example: '2021'
                        status:
                          type: string
                          example: ready
                        is_favorite:
                          type: boolean
                          example: false
                  links:
                    type: object
                    properties:
                      first:
                        type: string
                        example: '/?page=1'
                      last:
                        type: string
                        example: '/?page=1'
                      prev:
                        type: string
                        example: null
                      next:
                        type: string
                        example: null
                  meta:
                    type: object
                    properties:
                      current_page:
                        type: integer
                        example: 1
                      from:
                        type: integer
                        example: 1
                      last_page:
                        type: integer
                        example: 1
                      links:
                        type: array
                        example:
                          -
                            url: null
                            label: pagination.previous
                            active: false
                          -
                            url: '/?page=1'
                            label: '1'
                            active: true
                          -
                            url: null
                            label: pagination.next
                            active: false
                        items:
                          type: object
                          properties:
                            url:
                              type: string
                              example: null
                            label:
                              type: string
                              example: pagination.previous
                            active:
                              type: boolean
                              example: false
                      path:
                        type: string
                        example: /
                      per_page:
                        type: integer
                        example: 8
                      to:
                        type: integer
                        example: 2
                      total:
                        type: integer
                        example: 2
      tags:
        - Film
      security: []
    post:
      summary: 'POST /api/v1/films'
      operationId: pOSTapiv1films
      description: "\n        Добавление фильма в базу. Проставляется статус pending.\n        У фильмов с этим статусом обновляются данные посредством обращения раз в день к данным omdbApi.\n        Выполнять действие может только модератор."
      parameters: []
      responses:
        201:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: 'message: Фильм с imdb_id tt4523465 добавлен в очередь на создание'
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'Нет активной сессии'
                properties:
                  error:
                    type: string
                    example: 'Нет активной сессии'
        403:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'Нет доступа к дейсвтию!'
                properties:
                  error:
                    type: string
                    example: 'Нет доступа к дейсвтию!'
      tags:
        - Film
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 'Поле value не может превышать 255 символов.'
                  example: wzibrzeifwcgsqgmtsylmf
                posterImage:
                  type: string
                  description: "Ссылка должна вести на картинку: https://.../image.png;\n        Допустимые расширения расширения задаются в config(\"filesystems.img_extensions\")"
                  example: 'https://img.freepik.com/free-photo/view-of-3d-adorable-cat-with-fluffy-clouds_23-2151113419.jpg'
                previewImage:
                  type: string
                  description: "Ссылка должна вести на картинку: https://.../image.png;\n        Допустимые расширения расширения задаются в config(\"filesystems.img_extensions\")"
                  example: 'https://img.freepik.com/free-photo/view-of-3d-adorable-cat-with-fluffy-clouds_23-2151113419.jpg'
                backgroundImage:
                  type: string
                  description: "Ссылка должна вести на картинку: https://.../image.png;\n        Допустимые расширения расширения задаются в config(\"filesystems.img_extensions\")"
                  example: 'https://img.freepik.com/free-photo/view-of-3d-adorable-cat-with-fluffy-clouds_23-2151113419.jpg'
                backgroundColor:
                  type: string
                  description: ''
                  example: null
                videoLink:
                  type: string
                  description: 'Валидный url адрес'
                  example: 'https://www.youtube.com/watch?v=dQw4w9WgQ'
                previewVideoLink:
                  type: 'Валидный url адрес'
                  description: ''
                  example: 'https://www.youtube.com/watch?v=dQw4w9WgQ'
                description:
                  type: string
                  description: 'Поле value не может превышать 1000 символов.'
                  example: 'Eos commodi voluptatem alias.'
                director:
                  type: string
                  description: 'Поле value не может превышать 255 символов.'
                  example: zwhmdbtnfgxuabg
                starring:
                  type: string
                  description: 'Актеры в фильме. Заполняется в строку, можно заполнять через запятую'
                  example: 'Актер 1, Актер 2'
                runTime:
                  type: integer
                  description: 'Время длительности фильма'
                  example: 140
                genre:
                  type: string
                  description: 'Жанры фильма. Заполняется в строку, жанры можно заполнять через запятую'
                  example: 'комедия, семейный'
                released:
                  type: integer
                  description: 'Год выпуска фильма'
                  example: '2003 или 2003-2005'
                status:
                  type: string
                  description: 'Статусы фильмов'
                  example: moderate
                  enum:
                    - ready
                    - pending
                    - moderate
                imdb_id:
                  type: string
                  description: 'imdbId, который должен соответстовать регулярному выражению: /^tt\d{7,8}$/'
                  example: tt1111111
              required:
                - imdb_id
...
  /api/v1/user:
    get:
      summary: ''
      operationId: getApiV1User
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    id: 1
                    name: guest@guest.com
                    email: guest@guest.com
                    file: uploads/avatar_66f02fc2ae648.png
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      name:
                        type: string
                        example: guest@guest.com
                      email:
                        type: string
                        example: guest@guest.com
                      file:
                        type: string
                        example: uploads/avatar_66f02fc2ae648.png
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'Нет активной сессии'
                properties:
                  error:
                    type: string
                    example: 'Нет активной сессии'
        403:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'Нет доступа к дейсвтию!'
                properties:
                  error:
                    type: string
                    example: 'Нет доступа к дейсвтию!'
      tags:
        - User
    patch:
      summary: ''
      operationId: patchApiV1User
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Ваш профиль был обновлен'
                  data:
                    id: 1
                    name: guest@guest.com
                    email: guest@guest.com
                    file: uploads/avatar_66f02fc2ae648.png
                properties:
                  message:
                    type: string
                    example: 'Ваш профиль был обновлен'
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      name:
                        type: string
                        example: guest@guest.com
                      email:
                        type: string
                        example: guest@guest.com
                      file:
                        type: string
                        example: uploads/avatar_66f02fc2ae648.png
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'Нет активной сессии'
                properties:
                  error:
                    type: string
                    example: 'Нет активной сессии'
        403:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'Нет доступа к дейсвтию!'
                properties:
                  error:
                    type: string
                    example: 'Нет доступа к дейсвтию!'
      tags:
        - User
      requestBody:
        required: false
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 'Поле value не может превышать 255 символов.'
                  example: aovpgecegeknlpukf
                email:
                  type: string
                  description: 'Поле value должно быть действительным адресом электронной почты. Поле value не может превышать 255 символов.'
                  example: carmstrong@example.com
                password:
                  type: string
                  description: 'Поле value должно быть длиной минимум 8 символов.'
                  example: '@gS,U\xX&'
                file:
                  type: string
                  format: binary
                  description: 'Поле value поддерживает только изображения Поле value не может превышать 10240 килобайт.'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: null
security:
  -
    default: []